refactor: migrate to api-layer entity
