+++ src/api/config/types.ts
@@ -1,0 +1,11 @@
+// /////////////////////////////////////
+// Authentication
+// /////////////////////////////////////
+
+export type AuthUser = {
+  id: number;
+  roleId: number;
+  name: string;
+  adminAccess: boolean;
+  appAccess: boolean | null; // access from applications.
+};
