+++ src/api/utilities/sign.ts
@@ -1,6 +1,6 @@
 import jwt from 'jsonwebtoken';
-import { AuthUser } from '../../config/types.js';
 import { env } from '../../env.js';
+import { AuthUser } from '../config/types.js';
 
 export const sign = (user: AuthUser, expiresIn: number): string => {
   return jwt.sign(user, env.SECRET, {
